2025/mem0/requirements.txt
injector==0.22.0
mem0ai==1.0.1
python-dotenv==1.2.1
PyYAML==6.0.3
strands-agents[openai]==1.16.0
strands-agents-tools==0.2.15
